pyproject.toml
[tool.poetry]
name = "ywh"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
pymongo = "^4.7.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
redis = "^5.0.5"
schedule = "^1.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
